@@ -30,7 +30,8 @@
     "sqlite3": "^5.1.6",
     "swagger-jsdoc": "^6.2.8",
     "swagger-ui-express": "^5.0.1",
-    "vue": "^2.6.14",
-    "wstcp": "^1.0.6"
+    "vue": "^3.5.13",
+    "wstcp": "^1.0.6",
+    "yabooks-app": "github:Yabooks/yabooks-app"
   }
 }
