@@ -1,5 +1,24 @@
-Vue.component("LedgerTab",
+const LedgerTab = (
 {
+    props: [ "doc" ],
+
+    components: { LedgerTx },
+
+    data()
+    {
+        return {
+            price_type: "gross"
+        };
+    },
+
+    methods:
+    {
+        togglePriceType()
+        {
+            this.price_type = (this.price_type === "gross") ? "net" : "gross";
+        }
+    },
+
     template: `
         <div class="item">
             <h3>Invoice Items</h3>
@@ -10,7 +29,7 @@ Vue.component("LedgerTab",
                     <th @dblclick="togglePriceType()">Price {{ price_type }}</th>
                     <th>Tax</th>
                 </tr>
-                <tr v-for="(item, i) in (invoice.items || [])">
+                <tr v-for="(item, i) in (doc?.ledger_transactions || [])">
                     <td></td>
                     <td></td>
                     <td></td>
@@ -26,7 +45,7 @@ Vue.component("LedgerTab",
                     <th>Amount</th>
                     <th>Due</th>
                 </tr>
-                <tr v-for="(item, i) in (invoice.payments || [])">
+                <tr v-for="(item, i) in (doc?.payments || [])">
                     <td></td>
                     <td></td>
                     <td></td>
